Add squizlabs/php_codesniffer as a dev dependency
The squizlabs/php_codesniffer has been added to the dev dependencies in the composer.json file. A new configuration file, phpcs.xml.dist, has been created for it and its relevant entries were added to .gitignore and symfony.lock as well. This will help in detecting violations of a defined set of coding standards.